third_party/blink/web_tests/fast/css/dynamic-class-pseudo-elements-expected.txt

Change classes affecting pseudo elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(r1, '::before').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(r1, '::before').backgroundColor is green
PASS getComputedStyle(r2, '::after').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(r2, '::after').backgroundColor is green
PASS getComputedStyle(r3, '::first-line').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 9
PASS getComputedStyle(r3, '::first-line').backgroundColor is green
PASS getComputedStyle(r4, '::first-letter').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 9
PASS getComputedStyle(r4, '::first-letter').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

 < Background of 'X' should be green
Background of 'X' should be green > 
Background should be green
Background of first letter should be green
